requirements.txt
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
Pillow==10.0.0
numpy==1.24.3
scikit-learn==1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
opencv-python==4.8.0.74
tqdm>=4.60.0
gunicorn>=20.0.0
transformers==4.31.0